Clear listeners, hide icons when appropriate
Preferences are shown using view holders.

The previous code did not clear the listeners or hide the icons if necessary.

The practical upshot of this was that if you had two or more slider preferences, *and* they situated more than a screen's height apart, the viewholder from the first one would get reused.

And if the first one enabled icons then the second one would show them. And clicking on the second one would also call the listeners for the first one.